+++ src/main/kotlin/com/cluttered/cryptocurrency/model/Order.kt
@@ -1,8 +1,9 @@
 package com.cluttered.cryptocurrency.model
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class Order(
-        @SerializedName("Quantity") val quantity: Double,
-        @SerializedName("Rate") val rate: Double
+        @SerializedName("Quantity") val quantity: BigDecimal,
+        @SerializedName("Rate") val rate: BigDecimal
 )
